Limited bidirectional testing complete. Throughput with no routing hops, 8 slot windows, at CR=4/5 SF=7 BW=125k is about 250 bytes per second.  At 4 slot windows, it about 225.
Note to self: although we 'negotiate' a window size at connect time (the
CONNECT_ACK is sent with a 'sequence' number presenting the input window
size) it would be useful to be able to dynamically adjust the window during
a connection.  This suggests another packet type, treated as the CONNECT_ACK
where the size to use for an output_window so as to not overload the receiver.
At the moment, the CONNECT_ACK 'sequence' is use if it is smaller than the
compiled-in value.