Added lsocket packet cache to avoid reallocating a sending packets that are already on the radio transmit queue some where.
Added a packet_lock for each packet that uses a global mutex
to control and marks a locked packet with the thread id of the
locking thread.  It's not a  'blocking' lock - only returning
success or failure which is sufficient.

The radio transmit handler will attempt to lock a packet before
transmission and if fails, moves  it to the end of the queue.

The lsocket code which creates the control packets (connect/disconnect
etc.) maintains a cache of each stream cmd type and if it is attempting
to create a packet (for example CONNECT) it will first check the cache
for that socket and if found, will lock the packet and make the update
in place for any different flags, etc.

It will then 'route' the packet (which will ignore the actual deliver
to the transmit queue as the packet will already be 'transmitting...'.
When the packet is finally transmitted, the final transmitter will
thrown the 'has been transmitted' callback which will remove the
packet from the cache and release it.