Duty Cycle receive mode works but with some issues.  It seems the chip is sending an interrupt on each of the DutyCycle read cycles, even if the IRQ mask and DIO mask are off.  This is a significant bug and gives my driver headaches with all of the interrupt traffic that must be done: each IRQ needs a probe for IRQ status and DEVICE status (two SPI transactions) that end up going busy before completing, as the device goes back to the READ part of the cycle before the driver can get the second status message back.
Am researching solutions, but looking at the meshtastic drivers (which
are apparently by their hand) don't care and they never try to get
the data rate very high.

For now we will have to be contented with 125k bw and decent spreading
factors so we don't run up against this brick wall.